+++ changelog.md
@@ -6,6 +6,10 @@
 
 - [#4877](https://github.com/ignite/cli/pull/4877) Remove app config and ibc add route placeholders.
 
+### Changes
+
+- [#4878](https://github.com/ignite/cli/pull/4878) Improve the `xast` package readability.
+
 ## [`v29.8.0`](https://github.com/ignite/cli/releases/tag/v29.8.0)
 
 ## Features